Delete LOADEDMODULES cache
This cache is not used by the runtime and it is not compiled into debugger (mscordbi). It looks like a left-over from fusion.
