+++ src/coreclr/md/compiler/CMakeLists.txt
@@ -16,7 +16,6 @@ set(MDCOMPILER_SOURCES
   regmeta_emit.cpp
   regmeta_import.cpp
   regmeta_imetadatatables.cpp
-  regmeta_vm.cpp
 )
 
 set(MDCOMPILER_HEADERS
