+++ src/coreclr/md/compiler/disp.cpp
@@ -102,9 +102,6 @@ Disp::DefineScope(
     // Get the requested interface.
     IfFailGo(pMeta->QueryInterface(riid, (void **)ppIUnk));
 
-    // Add the new RegMeta to the cache.
-    IfFailGo(pMeta->AddToCache());
-
     LOG((LOGMD, "{%08x} Created new emit scope\n", pMeta));
 
 ErrExit:
@@ -182,32 +179,9 @@ Disp::OpenRawScope(
     _ASSERTE(!IsOfReserved(dwOpenFlags));
 #endif //!FEATURE_METADATA_LOAD_TRUSTED_IMAGES
 
-    {
-    }
-
     if (IsOfReadOnly(dwOpenFlags) && IsOfReadWrite(dwOpenFlags))
     {   // Invalid combination of flags - ofReadOnly & ofWrite
         IfFailGo(E_INVALIDARG);
-    }
-    // If open-for-read, and there is already an open-for-read copy, return it.
-    if (IsOfReadOnly(dwOpenFlags))
-    {
-        RegMeta::FindCachedReadOnlyEntry(szFileName, dwOpenFlags, &pMeta);
-        if (pMeta != NULL)
-        {
-            // Return the requested interface.
-            hr = pMeta->QueryInterface(riid, (void **) ppIUnk);
-            if (FAILED(hr))
-            {
-                pMeta = NULL; // Don't delete cached RegMeta!
-            }
-            else
-            {
-                pMeta->Release(); // Give back refcount from QI
-            }
-
-            goto ErrExit;
-        }
     }
     // Create a new coclass for this guy.
     pMeta = new (nothrow) RegMeta();
@@ -227,15 +201,6 @@ Disp::OpenRawScope(
 
     // Obtain the requested interface.
     IfFailGo(pMeta->QueryInterface(riid, (void **)ppIUnk) );
-
-    // Add the new RegMeta to the cache.  If this is read-only, any future opens will
-    //  find this entry.  If, due to another thread concurrently opening the same file,
-    //  there is already another copy in the cache, well, then there will be two
-    //  read-only copies in the cache.  This is considered to be somewhat of a corner
-    //  case, and the only harm is temporary memory usage.  All requests will be
-    //  satisfied by one or the other (depending on search algorithm), and eventually,
-    //  the "other" copy will be released.
-    IfFailGo(pMeta->AddToCache());
 
 #if defined(_DEBUG)
     if (CLRConfig::GetConfigValue(CLRConfig::INTERNAL_MD_RegMetaDump))
@@ -316,9 +281,6 @@ HRESULT Disp::OpenRawScopeOnMemory(        // Return code.
 
     // Return the requested interface.
     IfFailGo( pMeta->QueryInterface(riid, (void **) ppIUnk) );
-
-    // Add the new RegMeta to the cache.
-    IfFailGo(pMeta->AddToCache());
 
 #if defined(_DEBUG)
     if (CLRConfig::GetConfigValue(CLRConfig::INTERNAL_MD_RegMetaDump))
@@ -446,9 +408,6 @@ Disp::DefinePortablePdbScope(
     // Get the requested interface.
     IfFailGo(pMeta->QueryInterface(riid, (void**)ppIUnk));
 
-    // Add the new RegMeta to the cache.
-    IfFailGo(pMeta->AddToCache());
-
     LOG((LOGMD, "{%08x} Created new emit scope\n", pMeta));
 
 ErrExit:
@@ -527,9 +486,6 @@ HRESULT Disp::OpenRawScopeOnCustomDataSource(        // Return code.
 
     // Return the requested interface.
     IfFailGo(pMeta->QueryInterface(riid, (void **)ppIUnk));
-
-    // Add the new RegMeta to the cache.
-    IfFailGo(pMeta->AddToCache());
 
 #if defined(_DEBUG)
     if (CLRConfig::GetConfigValue(CLRConfig::INTERNAL_MD_RegMetaDump))
@@ -846,20 +802,6 @@ HRESULT Disp::GetOption(                // Return code.
 ErrExit:
     return hr;
 } // Disp::GetOption
-
-#if defined(FEATURE_METADATA_IN_VM)
-
-//---------------------------------------------------------------------------------------
-//
-// Process detach destruction.
-// Called from DllMain of clr.dll/RoMetadata.dll/MidlrtMd.dll.
-//
-void DeleteMetaData()
-{
-    LOADEDMODULES::DeleteStatics();
-}
-
-#endif //FEATURE_METADATA_IN_VM
 
 //
 // This is the entrypoint for usages of MetaData that need to start with the dispenser (e.g.
